Fix 32/64-bit bug in SparseBitArray and add tests
Correct Set/SetAll to use 64-bit masks (1ul/ulong.MaxValue) instead of 32-bit, ensuring proper bit operations across all 64 bits. Add thorough unit tests to verify correct behavior for lower/upper 32 bits, ulong boundaries, and SetAll.

+++ Source/Mosa.Compiler.Common/SparseBitArray.cs
@@ -72,12 +72,12 @@ public sealed class SparseBitArray
 
 		if (value)
 		{
-			Bits[index / 64] |= 1u << (index % 64);
+			Bits[index / 64] |= 1ul << (index % 64);
 		}
 		else
 
 		{
-			Bits[index / 64] &= ~(1u << (index % 64));
+			Bits[index / 64] &= ~(1ul << (index % 64));
 		}
 	}
 
@@ -85,7 +85,7 @@ public sealed class SparseBitArray
 	{
 		for (int i = 0; i < Bits.Length; i++)
 		{
-			Bits[i] = value ? uint.MaxValue : 0u;
+			Bits[i] = value ? ulong.MaxValue : 0u;
 		}
 	}
